+++ raphodo/argumentsparse.py
@@ -119,7 +119,10 @@ def get_parser(formatter_class=RawDescriptionHelpFormatter) -> ArgumentParser:
         type=str,
         metavar=_("PATH"),
         dest="this_computer_location",
-        help=_("The PATH on this computer from which to download."),
+        help=_(
+            "The PATH on this computer from which to download. If specified, turns on "
+            "downloading from this computer"
+        ),
     )
     parser.add_argument(
         "--photo-destination",